sync and populate db only if ENV=dev

+++ src/models/user.ts
@@ -1,7 +1,6 @@
 import bcrypt from "bcrypt";
 import { DataTypes } from "sequelize";
-import { getEnvVar } from "../helpers";
-import { sequelize } from "../helpers";
+import { getEnvVar, sequelize } from "../helpers";
 
 const User = sequelize.define(
   "User",
@@ -27,20 +26,22 @@ const User = sequelize.define(
 
 (async () => {
   try {
-    await User.sync({ force: true });
-    console.log("user table synced (forced)");
+    if (getEnvVar("ENV") === "dev") {
+      await User.sync({ force: true });
+      console.log("user table synced (forced)");
 
-    const email = getEnvVar("ADMIN_EMAIL");
-    const password = getEnvVar("ADMIN_PASSWORD");
-    const hashedPassword = await bcrypt.hash(
-      password,
-      parseInt(getEnvVar("BCRYPT_SALT"), 10),
-    );
-    await User.create({
-      email,
-      password: hashedPassword,
-    });
-    console.log("user table populaton succeeded");
+      const email = getEnvVar("ADMIN_EMAIL");
+      const password = getEnvVar("ADMIN_PASSWORD");
+      const hashedPassword = await bcrypt.hash(
+        password,
+        parseInt(getEnvVar("BCRYPT_SALT"), 10),
+      );
+      await User.create({
+        email,
+        password: hashedPassword,
+      });
+      console.log("user table populaton succeeded");
+    }
   } catch (error) {
     console.error(`user table sync/populate failed: ${error}`);
     process.exit(1);
